extract mocks to prevent duplicate tests
