+++ test/observableApiAuth-test.ts
@@ -1,11 +1,9 @@
 import assert from "node:assert";
 import {commandRequiresAuthenticationMessage} from "../src/commandInstruction.js";
 import {CliError} from "../src/error.js";
-import {type AuthEffects, login, logout, whoami} from "../src/observableApiAuth.js";
-import {TestClackEffects} from "./mocks/clack.js";
-import {MockLogger} from "./mocks/logger.js";
+import {login, logout, whoami} from "../src/observableApiAuth.js";
+import {MockAuthEffects} from "./mocks/authEffects.js";
 import {getCurrentObservableApi, mockObservableApi, validApiKey} from "./mocks/observableApi.js";
-import {MockConfigEffects} from "./observableApiConfig-test.js";
 
 describe("login command", () => {
   mockObservableApi();
@@ -109,53 +107,3 @@ describe("whoami command", () => {
     ]);
   });
 });
-
-class Deferred<T = unknown> {
-  promise: Promise<T>;
-  resolve!: (value: T) => void;
-  reject!: (reason?: any) => void;
-
-  constructor() {
-    this.promise = new Promise((resolve, reject) => {
-      this.resolve = resolve;
-      this.reject = reject;
-    });
-  }
-}
-
-export class MockAuthEffects extends MockConfigEffects implements AuthEffects {
-  public logger = new MockLogger();
-  public openBrowserDeferred = new Deferred<string>();
-  public setApiKeyDeferred = new Deferred<{id: string; apiKey: string}>();
-  public exitDeferred = new Deferred<void>();
-  public isTty: boolean;
-  public observableApiKey: string | null = null;
-  public outputColumns: number = 80;
-  public clack = new TestClackEffects();
-
-  constructor({apiKey = null, isTty = true}: {apiKey?: string | null; isTty?: boolean} = {}) {
-    super();
-    this.observableApiKey = apiKey;
-    this.isTty = isTty;
-  }
-
-  async getObservableApiKey(effects: AuthEffects = this) {
-    if (effects !== this) throw new Error("don't pass unrelated effects to mock effects methods");
-    if (!this.observableApiKey) return null;
-    return {source: "test" as const, key: this.observableApiKey};
-  }
-  async setObservableApiKey(info) {
-    this.setApiKeyDeferred.resolve(info);
-    this.observableApiKey = info?.key;
-  }
-  openUrlInBrowser(url: string) {
-    this.openBrowserDeferred.resolve(url);
-    return Promise.resolve(undefined);
-  }
-  waitForEnter() {
-    return Promise.resolve(undefined);
-  }
-  exitSuccess() {
-    this.exitDeferred.resolve(undefined);
-  }
-}
